+++ bud/src/main/cs/BudCli.cs
@@ -4,6 +4,7 @@ using System.Collections.Immutable;
 using System.IO;
 using Bud.Build;
 using Bud.Cli;
+using Bud.Cli.Macros;
 using Bud.Commander;
 using Bud.Info;
 using Bud.Util;
@@ -11,7 +12,7 @@ using CommandLine;
 
 namespace Bud {
   public static class BudCli {
-    private static readonly ImmutableList<Command> DefaultCommands = ImmutableList.Create<Command>(new KeyCommand(BuildKeys.Build.Id));
+    private static readonly ImmutableList<Command> DefaultCommands = ImmutableList.Create<Command>(new MacroCommand(InteractiveModeMacro.InteractiveModeMacroName));
 
     public static void Main(string[] args) {
       var cliArguments = new CliArguments();